+++ CHANGELOG.md
@@ -9,6 +9,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Warning box if composite data does not exist
 * Scaling for single color images
 * Undo of thematic map edits
+* Added colored theme background for theme radio buttons
+* Added prompting box for downloads
+* Font color for theme radio button changes depending on darkness of theme color, e.g. white for black backgrounds
+
+### Changed
+* Resized the thematic map annotator window
 
 ## [0.2.5]
 ### Added